feat: swap Flowbite data- attrs for custom trigger
This change requies defining a data-controller="flowbite--modal-target"
attribute for any modal that you want to, well, open like a modal!

It links the modal to the "stub" target controller that allows the
stimulus outlet system to connect the target to its trigger (and vice
versa). The bulk of the work happens in the trigger controller for both
a given component, and also the parent class that any of the component
trigger controllers inherit from.

Any component that defines a data-controller="flowbite--modal-trigger"
needs to also define an outlet that the trigger will "target", hence
connecting our modal to the "flowbite--modal-target" data-controller.

The outlet takes the form:

data-[controller-name]-[outlet-name]-outlet="[selector]"

The reason for the double-dashes in our outlet data attribute definition
is because we have the flowbite trigger and target data-controllers
nested in a "flowbite" folder inside
app/javascript/controllers/flowbite. The -- is the built-in
convention/requirement for specifying a controller in a nested folder.

Under the hood, stimulus is "connecting" to our trigger when that dom
node shows up in the document. It does this with a mutationObserver,
which ensures a reliable mechanism for hooking the trigger to the target
when either shows up in the DOM. The stimulus connect() method is the
key to establishing the link between the trigger and target in the
context of instantiating a new Flowbite component in JavaScript.

The beauty of stimulus using a mutationObserver pattern in the connect()
method is that the connection will happen on page load, or via any other
mechanism that results in a trigger/target being added to the DOM. For
example, when streaming in an updated trigger or target from a
turbo_stream in a resources controller like products_controller.rb in
this sample app.

Without taking this more fine-grained approach to Flowbite component
instantiation, the default flowbite.turbo.js variant will re-initialize
every Flowbite component in the DOM on every turbo:load request
regardless of whether it's a full page load or the smallest of
turbo_stream updates. This constant full re-initialization breaks key
components in various ways, but a big one is if you're interacting with
a drawer and creating new product records, for example, but you want the
drawer to stay open while you're adding all the products. With the
default flowbite.turbo.js approach, the drawer will close after every
form submission when the turbo_stream updating the DOM causes all of the
components to re-initialize. So to have more control over the Flowbite
JavaScript component lifecycle, we needed to pull in and work with the
Flowbite component ES modules more directly. In fact, this app doesn't
even include the flowbite.js file at all! We just work with the
components, directly, and set them up and tear them down on-demand.